+++ src/otx/__init__.py
@@ -1,9 +1,9 @@
 """OpenVINO Training Extensions."""
 
-# Copyright (C) 2024 Intel Corporation
+# Copyright (C) 2024-2025 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-__version__ = "2.3.0rc0"
+__version__ = "2.4.0dev"
 
 import os
 from pathlib import Path