+++ src/otx/engine/utils/auto_configurator.py
@@ -14,7 +14,7 @@ from warnings import warn
 import datumaro
 from jsonargparse import ArgumentParser, Namespace
 
-from otx.core.config.data import SamplerConfig, SubsetConfig, TileConfig, UnlabeledDataConfig, VisualPromptingConfig
+from otx.core.config.data import SamplerConfig, SubsetConfig, TileConfig, VisualPromptingConfig
 from otx.core.data.module import OTXDataModule
 from otx.core.model.base import OTXModel, OVModel
 from otx.core.types import PathLike
@@ -40,21 +40,18 @@ DEFAULT_CONFIG_PER_TASK = {
     OTXTaskType.ROTATED_DETECTION: RECIPE_PATH / "rotated_detection" / "maskrcnn_r50.yaml",
     OTXTaskType.SEMANTIC_SEGMENTATION: RECIPE_PATH / "semantic_segmentation" / "litehrnet_18.yaml",
     OTXTaskType.INSTANCE_SEGMENTATION: RECIPE_PATH / "instance_segmentation" / "maskrcnn_r50.yaml",
-    OTXTaskType.ACTION_CLASSIFICATION: RECIPE_PATH / "action_classification" / "x3d.yaml",
     OTXTaskType.ANOMALY: RECIPE_PATH / "anomaly" / "padim.yaml",
     OTXTaskType.ANOMALY_CLASSIFICATION: RECIPE_PATH / "anomaly_classification" / "padim.yaml",
     OTXTaskType.ANOMALY_SEGMENTATION: RECIPE_PATH / "anomaly_segmentation" / "padim.yaml",
     OTXTaskType.ANOMALY_DETECTION: RECIPE_PATH / "anomaly_detection" / "padim.yaml",
     OTXTaskType.VISUAL_PROMPTING: RECIPE_PATH / "visual_prompting" / "sam_tiny_vit.yaml",
-    OTXTaskType.ZERO_SHOT_VISUAL_PROMPTING: RECIPE_PATH / "zero_shot_visual_prompting" / "sam_tiny_vit.yaml",
     OTXTaskType.KEYPOINT_DETECTION: RECIPE_PATH / "keypoint_detection" / "rtmpose_tiny.yaml",
-    OTXTaskType.DIFFUSION: RECIPE_PATH / "diffusion" / "sd_huggingface.yaml",
-    OTXTaskType.OBJECT_DETECTION_3D: RECIPE_PATH / "object_detection_3d" / "monodetr3d.yaml",
 }
 
 TASK_PER_DATA_FORMAT = {
     "imagenet_with_subset_dirs": [OTXTaskType.MULTI_CLASS_CLS, OTXTaskType.H_LABEL_CLS],
     "datumaro": [OTXTaskType.MULTI_LABEL_CLS],
+    "coco_person_keypoints": [OTXTaskType.KEYPOINT_DETECTION],
     "coco_instances": [
         OTXTaskType.DETECTION,
         OTXTaskType.ROTATED_DETECTION,
@@ -67,9 +64,7 @@ TASK_PER_DATA_FORMAT = {
         OTXTaskType.INSTANCE_SEGMENTATION,
         OTXTaskType.VISUAL_PROMPTING,
     ],
-    "coco_captions": [OTXTaskType.DIFFUSION],
     "common_semantic_segmentation_with_subset_dirs": [OTXTaskType.SEMANTIC_SEGMENTATION],
-    "kinetics": [OTXTaskType.ACTION_CLASSIFICATION],
     "mvtec": [
         OTXTaskType.ANOMALY,
         OTXTaskType.ANOMALY_CLASSIFICATION,
@@ -87,14 +82,11 @@ OVMODEL_PER_TASK = {
     OTXTaskType.INSTANCE_SEGMENTATION: "otx.core.model.instance_segmentation.OVInstanceSegmentationModel",
     OTXTaskType.SEMANTIC_SEGMENTATION: "otx.core.model.segmentation.OVSegmentationModel",
     OTXTaskType.VISUAL_PROMPTING: "otx.core.model.visual_prompting.OVVisualPromptingModel",
-    OTXTaskType.ZERO_SHOT_VISUAL_PROMPTING: "otx.core.model.visual_prompting.OVZeroShotVisualPromptingModel",
-    OTXTaskType.ACTION_CLASSIFICATION: "otx.core.model.action_classification.OVActionClsModel",
     OTXTaskType.ANOMALY: "otx.algo.anomaly.openvino_model.AnomalyOpenVINO",
     OTXTaskType.ANOMALY_CLASSIFICATION: "otx.algo.anomaly.openvino_model.AnomalyOpenVINO",
     OTXTaskType.ANOMALY_DETECTION: "otx.algo.anomaly.openvino_model.AnomalyOpenVINO",
     OTXTaskType.ANOMALY_SEGMENTATION: "otx.algo.anomaly.openvino_model.AnomalyOpenVINO",
     OTXTaskType.KEYPOINT_DETECTION: "otx.core.model.keypoint_detection.OVKeypointDetectionModel",
-    OTXTaskType.OBJECT_DETECTION_3D: "otx.core.model.detection_3d.OV3DDetectionModel",
 }
 
 
@@ -233,7 +225,6 @@ class AutoConfigurator:
         train_config = data_config.pop("train_subset")
         val_config = data_config.pop("val_subset")
         test_config = data_config.pop("test_subset")
-        unlabeled_config = data_config.pop("unlabeled_subset", {})
         tile_config = data_config.pop("tile_config", {})
         vpm_config = data_config.pop("vpm_config", {})
 
@@ -248,10 +239,6 @@ class AutoConfigurator:
             train_subset=SubsetConfig(sampler=SamplerConfig(**train_config.pop("sampler", {})), **train_config),
             val_subset=SubsetConfig(sampler=SamplerConfig(**val_config.pop("sampler", {})), **val_config),
             test_subset=SubsetConfig(sampler=SamplerConfig(**test_config.pop("sampler", {})), **test_config),
-            unlabeled_subset=UnlabeledDataConfig(
-                sampler=SamplerConfig(**unlabeled_config.pop("sampler", {})),
-                **unlabeled_config,
-            ),
             tile_config=TileConfig(**tile_config),
             vpm_config=VisualPromptingConfig(**vpm_config),
             **data_config,
@@ -422,7 +409,6 @@ class AutoConfigurator:
         subset_config.to_tv_image = ov_config[f"{subset}_subset"]["to_tv_image"]
         datamodule.image_color_channel = ov_config["image_color_channel"]
         datamodule.tile_config.enable_tiler = False
-        datamodule.unlabeled_subset.data_root = None
         msg = (
             f"For OpenVINO IR models, Update the following {subset} \n"
             f"\t transforms: {subset_config.transforms} \n"
@@ -439,7 +425,6 @@ class AutoConfigurator:
             train_subset=datamodule.train_subset,
             val_subset=datamodule.val_subset,
             test_subset=datamodule.test_subset,
-            unlabeled_subset=datamodule.unlabeled_subset,
             tile_config=datamodule.tile_config,
             vpm_config=datamodule.vpm_config,
             image_color_channel=datamodule.image_color_channel,