+++ tests/integration/api/test_engine_api.py
@@ -34,8 +34,6 @@ def test_engine_from_config(
     """
     if task not in DEFAULT_CONFIG_PER_TASK:
         pytest.skip("Only the Task has Default config is tested to reduce unnecessary resources.")
-    if task.lower() in ("action_classification"):
-        pytest.xfail(reason="xFail until this root cause is resolved on the Datumaro side.")
     if task.lower() in ("h_label_cls"):
         pytest.skip(
             reason="H-labels require num_multiclass_head, num_multilabel_classes, which skip until we have the ability to automate this.",
@@ -48,19 +46,14 @@ def test_engine_from_config(
         work_dir=tmp_path_train,
         device=fxt_accelerator,
     )
-    if task.lower() == "zero_shot_visual_prompting":
-        engine.model.infer_reference_info_root = Path(tmp_path_train)
-        # update litmodule.hparams to reflect changed hparams
-        engine.model.hparams.update({"infer_reference_info_root": str(engine.model.infer_reference_info_root)})
 
     # Check OTXModel & OTXDataModule
     assert isinstance(engine.model, OTXModel)
     assert isinstance(engine.datamodule, OTXDataModule)
 
-    max_epochs = 2 if task.lower() != "zero_shot_visual_prompting" else 1
+    max_epochs = 2
     train_metric = engine.train(max_epochs=max_epochs)
-    if task.lower() != "zero_shot_visual_prompting":
-        assert len(train_metric) > 0
+    assert len(train_metric) > 0
 
     test_metric = engine.test()
     assert len(test_metric) > 0
@@ -71,7 +64,6 @@ def test_engine_from_config(
     # A Task that doesn't have Export implemented yet.
     # [TODO]: Enable should progress for all Tasks.
     if task in [
-        OTXTaskType.ACTION_CLASSIFICATION,
         OTXTaskType.H_LABEL_CLS,
         OTXTaskType.ROTATED_DETECTION,
         OTXTaskType.ANOMALY,
@@ -93,15 +85,7 @@ def test_engine_from_config(
 
     # Test with IR Model
     if task in OVMODEL_PER_TASK:
-        if task.lower() in ["visual_prompting", "zero_shot_visual_prompting"]:
-            if task.lower() == "zero_shot_visual_prompting":
-                engine.model = engine._auto_configurator.get_ov_model(
-                    model_name=str(exported_model_path["decoder"]),
-                    label_info=engine.datamodule.label_info,
-                )
-                engine.model.infer_reference_info_root = Path(tmp_path_train)
-                # update litmodule.hparams to reflect changed hparams
-                engine.model.hparams.update({"infer_reference_info_root": str(engine.model.infer_reference_info_root)})
+        if task.lower() == "visual_prompting":
             test_metric_from_ov_model = engine.test(checkpoint=exported_model_path["decoder"], accelerator="cpu")
         else:
             test_metric_from_ov_model = engine.test(checkpoint=exported_model_path, accelerator="cpu")