+++ tests/integration/cli/test_cli.py
@@ -30,8 +30,7 @@ def fxt_trained_model(
     recipe = request.param
     recipe_split = recipe.split("/")
     model_name = recipe_split[-1].split(".")[0]
-    is_semisl = model_name.endswith("_semisl")
-    task = recipe_split[-2] if not is_semisl else recipe_split[-3]
+    task = recipe_split[-2]
 
     # 1) otx train
     tmp_path_train = tmp_path / f"otx_train_{model_name}"
@@ -47,17 +46,9 @@ def fxt_trained_model(
         "--engine.device",
         fxt_accelerator,
         "--max_epochs",
-        "1" if task in ("zero_shot_visual_prompting") else "2",
+        "2",
         *fxt_cli_override_command_per_task[task],
     ]
-
-    if is_semisl:
-        command_cfg.extend(
-            [
-                "--data.unlabeled_subset.data_root",
-                fxt_target_dataset_per_task[f"{task}_semisl"],
-            ],
-        )
 
     run_main(command_cfg=command_cfg, open_subprocess=fxt_open_subprocess)
 
@@ -125,15 +116,6 @@ def test_otx_e2e(
         "--checkpoint",
         str(ckpt_file),
     ]
-    # Zero-shot visual prompting needs to specify `infer_reference_info_root`
-    if task in ["zero_shot_visual_prompting"]:
-        idx_task = str(ckpt_file).split("/").index(f"otx_train_{model_name}")
-        command_cfg.extend(
-            [
-                "--model.init_args.infer_reference_info_root",
-                str(ckpt_file.parents[-idx_task] / f"otx_train_{model_name}/outputs/.latest/train"),
-            ],
-        )
 
     run_main(command_cfg=command_cfg, open_subprocess=fxt_open_subprocess)
 
@@ -153,15 +135,11 @@ def test_otx_e2e(
         ]
     ):
         return
-    if task in ("visual_prompting", "zero_shot_visual_prompting"):
+    if task == "visual_prompting":
         fxt_export_list = [
             ExportCase2Test("ONNX", False, "exported_model_decoder.onnx"),
             ExportCase2Test("OPENVINO", False, "exported_model_decoder.xml"),
         ]  # TODO (sungchul): EXPORTABLE_CODE will be supported
-
-    if task == "object_detection_3d":
-        # exportable code and demo package are not supported for OD 3D
-        fxt_export_list.pop(-1)
 
     overrides = fxt_cli_override_command_per_task[task]
 
@@ -202,7 +180,7 @@ def test_otx_e2e(
         msg = "There is no OV IR."
         raise RuntimeError(msg)
     exported_model_path = str(ov_files[0])
-    if task in ("visual_prompting", "zero_shot_visual_prompting"):
+    if task == "visual_prompting":
         recipe = str(Path(recipe).parents[0] / "openvino_model.yaml")
 
     overrides = fxt_cli_override_command_per_task[task]
@@ -224,15 +202,6 @@ def test_otx_e2e(
         "--checkpoint",
         exported_model_path,
     ]
-    # Zero-shot visual prompting needs to specify `infer_reference_info_root`
-    if task in ["zero_shot_visual_prompting"]:
-        idx_task = str(ckpt_file).split("/").index(f"otx_train_{model_name}")
-        command_cfg.extend(
-            [
-                "--model.init_args.infer_reference_info_root",
-                str(ckpt_file.parents[-idx_task] / f"otx_train_{model_name}/outputs/.latest/train"),
-            ],
-        )
 
     run_main(command_cfg=command_cfg, open_subprocess=fxt_open_subprocess)
 
@@ -257,10 +226,6 @@ def test_otx_e2e(
 
     if "keypoint" in recipe:
         print("Explain is not supported for keypoint detection")
-        return
-
-    if "monodetr3d" in recipe:
-        print("Explain is not supported for object detection 3d")
         return
 
     tmp_path_test = tmp_path / f"otx_export_xai_{model_name}"
@@ -402,12 +367,7 @@ def test_otx_ov_test(
         "instance_segmentation",
         "h_label_cls",
         "visual_prompting",
-        "zero_shot_visual_prompting",
         "anomaly",
-        "anomaly_classification",
-        "anomaly_detection",
-        "anomaly_segmentation",
-        "action_classification",
     ]:
         # OMZ doesn't have proper model for Pytorch MaskRCNN interface
         # TODO(Kirill):  Need to change this test when export enabled
@@ -469,8 +429,6 @@ def test_otx_adaptive_bs_e2e(
         pytest.skip("Adaptive batch size only supports GPU and XPU.")
     if task not in DEFAULT_CONFIG_PER_TASK:
         pytest.skip(f"Task {task} is not supported in the auto-configuration.")
-    if task == OTXTaskType.ZERO_SHOT_VISUAL_PROMPTING:
-        pytest.skip("ZERO_SHOT_VISUAL_PROMPTING doesn't support adaptive batch size.")
 
     task = task.lower()
     tmp_path_adap_bs = tmp_path / f"otx_adaptive_bs_{task}"
@@ -518,8 +476,6 @@ def test_otx_configurable_input_size_e2e(
     """
     if task not in DEFAULT_CONFIG_PER_TASK:
         pytest.skip(f"Task {task} is not supported in the auto-configuration.")
-    if task == OTXTaskType.ZERO_SHOT_VISUAL_PROMPTING:
-        pytest.skip(f"{task} doesn't support configurable input size.")
     if task == OTXTaskType.KEYPOINT_DETECTION:
         pytest.skip(f"{task} doesn't support configurable input size.")